Solution to Check Profit or Loss

+++ Check Profit or Loss@@6788dd63490637cfc44e1b98/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main
 {
-    int costprice,sellingprice;
-    scanf("%d %d",&costprice,&sellingprice);
+    float costprice,sellingprice;
+    scanf("%f %f",&costprice,&sellingprice);
     if(costprice>sellingprice)
     {
         printf("Loss");
